Implementation of a topology subgraph finding function, which can find any subgraph existing in a topology.
The unit test takes in the NSF backbone topology as the substrate and a request of 
4 node ring, both in NDL.
The output is a dot file (nsf-current.dot) which can be shown in GraphViz.
It showns all the three 4-node ring subgraphs marked by different colors. 
There was a python code doing the graph, but here the Java code directly outputs the .dot file.