Fixes ticket #207.
Introduced support for timeouts for RPCs. Right now only claim and query use timeouts. 
The other RPCs are designed to keep retrying after a network error and for now do not use a timeout while waiting for the incoming updateTicket/updateLease. They do respect the reservation's term, though and retries will stop once the reservation expires.

The timeouts for query and claim are both 2 minutes and can be adjusted by editing RPCManager.java.
To support RPC timeouts, I added two new facilities:
 - Globals.Timer is a global java.util.Timer that can be used to schedule timer tasks. This one is used by Query.
 - OrcaTimer is a singleton utility class that queues a task to an actor's timer queue when a timeout fires. i
   I added a new queue per actor to store fired timers. Timers can also be cancelled.

When claim times out, the reservation on the broker side is failed. The reservation can then be removed and the claim can be reissued by the operator.
When query times out this is treated the same way as an RPC failure and the corresponding notifications are delivered.

NOTE: axis seems to time out after 60 seconds if it cannot establish a connection with the remote end. The timeouts I introduced are set once for an RPC, regardless of how many times it is retried.